update dependencies
i'm guessing it's because this project was created with a version of npm from 3 years ago, but this wouldn't compile properly on my machine. the setup process made me feel as if it _had_ worked, despite the errors, but the public app.css and app.js were committed to the repo, so it was a false positive.