feature: form validation
Here, I've put validation directly in the controller to address the issues we're looking to solve. I'd love to refactor this to a Form Request class, if there's time... and maybe naming or adding a custom message for that regex would make things clearer for the reader.
I've updated the front end to display error message using Bootstrap 4.