bugfix: date/time display on list of Bookings
Here, I've opted to handle this from the server side - as handling this in Javascript has a few big drawbacks:
 - it's possible using native JS functions, but I hate the syntax
 - it's possible using Moment, but that's a big ol' library for not much
 - if the list is long, we're making the client side do a lot of work

The backend is made for processing data... and doing it this way doesn't really change the memory consumption (via debugbar)... so I'm sticking with it.

I could have added a cast to both "start" and "end", but I'm not a fan of this approach because there's bound to be a time/scenario where you'd like to alter this behaviour down the line.